:truck: 상세페이지 모듈화
