+++ src/routes/LocationDetail/constants.js
@@ -1,0 +1,30 @@
+export const OPTION_LIST = {
+  toiletType: [
+    { label: '공중화장실', value: '공중화장실' },
+    { label: '개방화장실', value: '개방화장실' },
+  ],
+  babyYn: [
+    { label: '가능', value: 'Y' },
+    { label: '불가능', value: 'N' },
+  ],
+  unusualYn: [
+    { label: '가능', value: 'Y' },
+    { label: '불가능', value: 'N' },
+  ],
+  cctvYn: [
+    { label: '있음', value: 'Y' },
+    { label: '없음', value: 'N' },
+  ],
+  alarmYn: [
+    { label: '있음', value: 'Y' },
+    { label: '없음', value: 'N' },
+  ],
+  pwdYn: [
+    { label: '있음', value: 'Y' },
+    { label: '없음', value: 'N' },
+  ],
+  openYn: [
+    { label: '예', value: 'Y' },
+    { label: '아니오', value: 'N' },
+  ],
+};
