+++ src/routes/LocationDetail/index.js
@@ -1,0 +1,1 @@
+export * from './LocationDetailPage';
